This is an analysis of the black Friday sales.
This is a black Friday sales dataset that depict the purchase that was made by customers. In the dataset we have more people within the age bracket of 26-35 that participated in the black Friday sales. Another insight from the dataset was that people within the age bracket of 51-55 made more purchase and male also made more purchase as compared to female. Some columns had outliers but it was treated in the dataset. From the dataset, stay in current city did not really determine the purchasing power of clients. The dataset was cleaned we carried out exploratory data analysis and finally we made predictions.